Fixed buffering limits.

+++ SpaceDropCXX/sd.h
@@ -48,7 +48,7 @@ enum State {
 
 State state = S_IDLE;
 
-const int BUFFERING_LIMIT = 100000000;
+const size_t BUFFERING_LIMIT = 100000000;
 
 const int BLOCK_SIZE = 65527;
 const int BLOCK_SIZE_UPDATE = BLOCK_SIZE * 4;
